ai_service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
celery==5.4.0
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
gunicorn==25.1.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
langchain==0.3.27
langchain-core==0.3.79
langgraph==0.2.67
packaging==25.0
PyJWT==2.10.1
pydantic==2.12.5
pydantic_core==2.41.5
qdrant-client==1.15.1
redis==5.2.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
prometheus-fastapi-instrumentator==7.1.0
